코드리뷰 정리

+++ src/main/java/com/pjw/retry_view/service/EventService.java
@@ -9,10 +9,9 @@ import com.pjw.retry_view.repository.EventRepository;
 import com.pjw.retry_view.request.WriteEventRequest;
 import jakarta.transaction.Transactional;
 import org.springframework.stereotype.Service;
+import org.springframework.util.CollectionUtils;
 
-import java.time.ZonedDateTime;
 import java.util.ArrayList;
-import java.util.Arrays;
 import java.util.List;
 import java.util.Objects;
 
@@ -34,9 +33,10 @@ public class EventService {
         return eventRepository.findById(id).orElseThrow(ResourceNotFoundException::new).toDTO();
     }
 
+    @Transactional
     public EventDTO saveEvent(WriteEventRequest req){
-        Event event = Event.newEventFromReq(req);
-        List<EventImage> images = req.getImages().stream().map(img -> EventImage.getEventImage(img.getId(), img.getImageUrl(), req.getCreatedBy())).toList();
+        Event event = Event.newOne(req.getContent(), req.getStartAt(), req.getEndAt(), req.getCreatedBy());
+        List<EventImage> images = req.getImages().stream().map(img -> EventImage.newOne(img.getId(), img.getImageUrl(), req.getCreatedBy())).toList();
         eventRepository.save(event);
 
         for(EventImage eventImage : images){
@@ -49,13 +49,15 @@ public class EventService {
     @Transactional
     public EventDTO updateEvent(WriteEventRequest req){
         Event event = eventRepository.findById(req.getId()).orElseThrow(ResourceNotFoundException::new);
-        List<EventImage> reqImages = req.getImages().stream().map(img -> EventImage.getEventImage(img.getId(), img.getImageUrl(), req.getCreatedBy())).toList();
+        List<EventImage> reqImages = req.getImages().stream().map(img -> EventImage.newOne(img.getId(), img.getImageUrl(), req.getCreatedBy())).toList();
 
         List<Long> imageIds = req.getImages().stream().map(WriteEventRequest.Image::getId).filter(Objects::nonNull).toList();
         List<Long> oldImageIds = event.getEventImage().stream().map(EventImage::getId).toList();
         List<Long> deleteImageIds = getDeleteImageIds(imageIds, oldImageIds);
-        eventImageRepository.deleteByIds(deleteImageIds);
-        event.getEventImage().removeIf(img->deleteImageIds.contains(img.getId()));
+        if(!CollectionUtils.isEmpty(deleteImageIds)) {
+            eventImageRepository.deleteByIds(deleteImageIds);
+            event.getEventImage().removeIf(img->deleteImageIds.contains(img.getId()));
+        }
 
         for(EventImage eventImage : reqImages){
             if(eventImage.getId() == null) {
@@ -64,10 +66,11 @@ public class EventService {
             }
         }
 
-        event.updateEvent(req);
+        event.updateEvent(req.getContent(), req.getStartAt(), req.getEndAt(), req.getUpdatedBy());
         return eventRepository.save(event).toDTO();
     }
 
+    @Transactional
     public void deleteEvent(Long id){
         Event event = eventRepository.findById(id).orElseThrow(ResourceNotFoundException::new);
         eventImageRepository.deleteByIds(event.getEventImage().stream().map(EventImage::getId).toList());
@@ -75,6 +78,7 @@ public class EventService {
     }
 
     public List<Long> getDeleteImageIds(List<Long> images, List<Long> oldImageIds){
+        if(CollectionUtils.isEmpty(images) || CollectionUtils.isEmpty(oldImageIds)) return null;
         List<Long> deleteImageIds = new ArrayList<>();
         for(Long id : oldImageIds){
             if(!images.contains(id)){
